backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.22.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
google-cloud-bigquery==3.13.0
pydantic==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.3.2
